packages/aero-assistant/training/requirements.txt
torch==2.1.2
transformers==4.37.0
datasets==2.16.1
peft==0.7.1
bitsandbytes==0.41.3
accelerate==0.26.1
sentencepiece==0.1.99
protobuf==4.25.2
tensorboard==2.15.1
wandb==0.16.2
scipy==1.12.0
scikit-learn==1.4.0
tqdm==4.66.1
pyyaml==6.0.1
pandas==2.1.4
numpy==1.26.3
evaluate==0.4.1
rouge-score==0.1.2
nltk==3.8.1
